Revert "Feature/details api service"
