+++ src/components/details/Tabs.tsx
@@ -1,28 +1,38 @@
 /** @jsxImportSource @emotion/react */
 
-import React from "react";
 import { css } from "@emotion/react";
 import media from "lib/styles/media";
 import palette from "lib/palette";
-import { tabListType } from "./Tab";
-import TabListView from "./TabListView";
+import React, { useState } from "react";
+import TabList from "./TabList";
 
-type Props = {
-  tabElements: tabListType[];
-  active: string;
-  currentContent: tabListType;
-  onChangeTab: (e: React.MouseEvent<HTMLOListElement>) => void;
+type tabListType = {
+  title: string;
+  component: React.ReactNode;
 };
 
-const TabView = ({ tabElements, active, currentContent, onChangeTab }: Props) => {
+type Props = {
+  tabList: tabListType[];
+};
+
+const Tabs = ({ tabList }: Props) => {
+  const [active, setActive] = useState(tabList[0].title);
+  const content = tabList.find((tab) => active === tab.title);
+
+  const handleChangeTabActive = (e: React.MouseEvent<HTMLOListElement>) => {
+    const target = e.target as HTMLElement;
+    if (target.tagName === "OL") return;
+    setActive(target.id);
+  };
+
   return (
     <div css={tabs}>
-      <ol className="tab-list" onClick={onChangeTab}>
-        {tabElements.map((tab, index) => (
-          <TabListView key={`${tab.title}_${index}`} title={tab.title} active={active} />
+      <ol className="tab-list" onClick={handleChangeTabActive}>
+        {tabList.map((tab, index) => (
+          <TabList key={`${tab.title}_${index}`} title={tab.title} active={active} />
         ))}
       </ol>
-      <div className="tab-content">{currentContent.component}</div>
+      <div className="tab-content">{content?.component}</div>
     </div>
   );
 };
@@ -69,4 +79,4 @@ const tabs = css`
   }
 `;
 
-export default TabView;
+export default Tabs;
